+++ CHANGELOG.rst
@@ -25,13 +25,13 @@ The format is based on
 and this project adheres to
 `Semantic Versioning <https://semver.org/spec/v2.0.0.html>`_.
 
-Develop
-*******
+Version 1.1.0 (June 2023)
+*************************
 
 Added
 -----
 
-- An argument ``trend_type`` of type :attr:`~.OTGaussianProcessRegressor.TREND_TYPE`` to :class:`.OTGaussianProcessRegressor; the trend type of the Gaussian process regressor can be either constant, linear or quadratic.
+- An argument ``trend_type`` of type :attr:`~.OTGaussianProcessRegressor.TREND_TYPE`` to :class:`.OTGaussianProcessRegressor`; the trend type of the Gaussian process regressor can be either constant, linear or quadratic.
 - A new optimization library :class:`SurrogateBasedOptimization` to perform EGO-like surrogate-based optimization on unconstrained problems.
 
 Fixed